+++ ScoreTracker/ScoreTracker.PlayerProgress/Infrastructure/Entities/PlayerStatsEntity.cs
@@ -10,16 +10,20 @@ namespace ScoreTracker.PlayerProgress.Infrastructure.Entities
         public int TotalRating { get; set; }
         public int HighestLevel { get; set; }
         public int ClearCount { get; set; }
-        public int SkillRating { get; set; }
+
+        // The four PUMBILITY pools are stored unrounded. A pool is fifty per-chart values that
+        // each carry a real fraction, so rounding one of them at rest discards precision the
+        // presentation layer is the only thing entitled to spend (docs/UX-GUIDELINES.md).
+        public double SkillRating { get; set; }
         public double AverageSkillLevel { get; set; }
         public int AverageSkillScore { get; set; }
-        public int SinglesRating { get; set; }
+        public double SinglesRating { get; set; }
         public double AverageSinglesLevel { get; set; }
         public int AverageSinglesScore { get; set; }
-        public int DoublesRating { get; set; }
+        public double DoublesRating { get; set; }
         public double AverageDoublesLevel { get; set; }
         public int AverageDoublesScore { get; set; }
-        public int CoOpRating { get; set; }
+        public double CoOpRating { get; set; }
         public int AverageCoOpScore { get; set; }
         public double CompetitiveLevel { get; set; }
         public double SinglesCompetitiveLevel { get; set; }
